requirements.txt
Flask==1.1.2
Flask-Script==2.0.5
gunicorn==19.9.0
eventlet==0.30.2
requests==2.28.1
jinja2<3.1.0
itsdangerous==2.0.1
werkzeug==2.0.3
SPARQLWrapper==2.0
opensearch-py==2.0.0
folium
pillow